backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
pymysql==1.1.0
cryptography==42.0.5
pydantic==2.6.4
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
